+++ packages/index-page/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.11-rc.1](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/index-page@1.5.11-rc.0...@twreporter/index-page@1.5.11-rc.1) (2024-03-27)
+
+**Note:** Version bump only for package @twreporter/index-page
+
+
+
+
+
+## [1.5.11-rc.0](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/index-page@1.5.10...@twreporter/index-page@1.5.11-rc.0) (2024-03-25)
+
+**Note:** Version bump only for package @twreporter/index-page
+
+
+
+
+
 ## [1.5.10](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/index-page@1.5.10-rc.5...@twreporter/index-page@1.5.10) (2024-03-13)
 
 **Note:** Version bump only for package @twreporter/index-page